Servant.Validate

Description

A package with "test suites" to help verify that your servant APIs are valid at compile-time. Currently the only property tested is that there are no duplicated paths. See README for more information on usage.

Documentation

class HasApiTree (api :: Type) where

Has a valid well-formed API Tree

Associated Types

type ToApiTree api :: ApiTree

Methods

sApiTree :: SApiTree (ToApiTree api)

Useful runtime witness of the API tree; use to inspect it with reflectApiTree. This is not used in any part of the actual validation; is just an extra treat.

class MethodString k

A type-level version of ReflectMethod.

Associated Types

type ToMethodString (x :: k) :: Symbol

validApiTree :: forall api. (HasApiTree api, Typeable (ToApiTree api)) => Proxy api -> ValidApiTree api

The full validator. To use:

serverApi :: Proxy ServerApi
serverApi = Proxy

validServerApi :: ValidApiTree ServerApi
validServerApi = validApiTree serverApi

type ValidApiTree api = TypeRep (ToApiTree api)

To be used with validApiTree.

reflectApiTree :: forall api. (HasApiTree api, Typeable (ToApiTree api)) => ApiTreeMap

Useful utility to view the routing structure of a tree; similar to layout.

reflectApiTree_ :: TypeRep (apiTree :: ApiTree) -> ApiTreeMap

Version of reflectApiTree taking an explicit TypeRep.

data SApiTree :: ApiTree -> Type where

Constructors

SBranch :: Prod SSym ms -> Prod (Tup SSym SApiTree) ts -> SApiTree ('Branch ms ts)

reflectSApiTree :: SApiTree api -> ApiTreeMap